+++ WalletWasabi/Extensions/VerboseBlockInfo.cs
@@ -10,25 +10,4 @@ namespace NBitcoin.RPC
 		public ulong Height { get; set; }
 		public List<VerboseTransactionInfo> Transactions { get; set; } = new List<VerboseTransactionInfo>();
 	}
-
-	public class VerboseTransactionInfo
-	{
-		public uint256 Id { get; set; }
-
-		public List<VerboseInputInfo> Inputs { get; set; } = new List<VerboseInputInfo>();
-		public List<VerboseOutputInfo> Outputs { get; set; } = new List<VerboseOutputInfo>();
-
-	}
-
-	public class VerboseInputInfo
-	{
-		public OutPoint OutPoint { get; set; }
-		public VerboseOutputInfo PrevOutput { get; set; }
-	}
-
-	public class VerboseOutputInfo
-	{
-		public Money Value { get; set; }
-		public Script ScriptPubKey { get; set; }
-	}
 }
