+++ WalletWasabi/Extensions/VerboseInputInfo.cs
@@ -1,0 +1,8 @@
+namespace NBitcoin.RPC
+{
+	public class VerboseInputInfo
+	{
+		public OutPoint OutPoint { get; set; }
+		public VerboseOutputInfo PrevOutput { get; set; }
+	}
+}
