+++ WalletWasabi/Extensions/VerboseOutputInfo.cs
@@ -1,0 +1,8 @@
+namespace NBitcoin.RPC
+{
+	public class VerboseOutputInfo
+	{
+		public Money Value { get; set; }
+		public Script ScriptPubKey { get; set; }
+	}
+}
